+++ include/aligator/core/callback-base.hpp
@@ -10,8 +10,6 @@ template <typename Scalar> struct CallbackBaseTpl {
   using Workspace = WorkspaceBaseTpl<Scalar>;
   using Results = ResultsBaseTpl<Scalar>;
   virtual void call(const Workspace &, const Results &) = 0;
-  /// Call this after linesearch.
-  virtual void post_linesearch_call(boost::any) {}
   virtual ~CallbackBaseTpl() = default;
 };