+++ include/aligator/solvers/fddp/solver-fddp.hxx
@@ -36,16 +36,17 @@ void SolverFDDPTpl<Scalar>::setup(const Problem &problem) {
       idx_where_constraints.push_back(i);
   }
   if (idx_where_constraints.size() > 0) {
-    ALIGATOR_FDDP_WARNING(
-        fmt::format("problem stages [{}] have constraints, "
-                    "which this solver cannot handle. "
-                    "Please use a penalized cost formulation.\n",
-                    fmt::join(idx_where_constraints, ", ")));
+    ALIGATOR_WARNING("SolverFDDP",
+                     fmt::format("problem stages [{}] have constraints, "
+                                 "which this solver cannot handle. "
+                                 "Please use a penalized cost formulation.\n",
+                                 fmt::join(idx_where_constraints, ", ")));
   }
   if (!problem.term_cstrs_.empty()) {
-    ALIGATOR_FDDP_WARNING("problem has at least one terminal constraint, which "
-                          "this solver cannot "
-                          "handle.\n");
+    ALIGATOR_WARNING("SolverFDDP",
+                     "problem has at least one terminal constraint, which "
+                     "this solver cannot "
+                     "handle.\n");
   }
 }
 
@@ -300,7 +301,7 @@ bool SolverFDDPTpl<Scalar>::run(const Problem &problem,
   logger.addColumn(BASIC_KEYS[5]);
   logger.addColumn(BASIC_KEYS[6]);
   logger.addColumn(BASIC_KEYS[7]);
-  logger.addColumn(BASIC_KEYS[8]);
+  logger.addColumn(BASIC_KEYS[9]);
   logger.printHeadline();
 
   // in Crocoddyl, linesearch xs is primed to use problem x0